Don't ignore CPPFLAGS on compilation
CPPFLAGS is part of the built-in rules for compilation. Although it is not
used by ipv6-toolkit itself, distros may pass flags on CPPFLAGS too to
comply with some policies.

Being part of the built-in rules, custom rules like used in the Makefile
should explicitly include CPPFLAGS. See:

http://en.wikipedia.org/wiki/CFLAGS
http://stackoverflow.com/questions/2754966/cflags-vs-cppflags

In the case of Debian, debuild passes flags for executable hardening.

http://lintian.debian.org/tags/hardening-no-fortify-functions.html