Replaced most hyphens for minus signs in man pages
In a man page, hyphen and minus are different signs and can be rendered
differently depending on the output. This can cause examples to fail
for users that copy and paste examples from rendered man pages.

Hyphen is coded like this: -
Minus is coded like this: \-
Options like --help use minus, like this: \-\-help.

This was detected by Debian lintian. See:

http://lintian.debian.org/tags/hyphen-used-as-minus-sign.html