Added unit testing for libipv6; only one test for now.
This was used to make sure we didn't break a modified function in
the future. Let this commit work as an example for the unit testing
technique.

Make will try to create a timestamp file "libipv6.ts" if execution of
"libipv6.t" succeeds. libipv6.t will be created from libipv6.t.c, the
unit testing code.

Using a timestamp makes it easy to let Make know if it is necessary to
run the test again or not. If the timestamp technique is not used, Make
will run the test unnecessarily on each call.