scan6: Endianness fix for ip6_plen
When ip6_plen was used for reading, the possible endianness mismatch
between host and network was not considered. (It was considered when
writing to ip6_plen). This fixes all occurrences thereof in scan6.c.

The issue was likely to go unnoticed because at typical sizes, the bug
would over-report IPv6 sizes, and because the end of the Ethernet frame
typically matches the end of the IPv6 frame anyway, unless odd or buggy
network hardware us used.