update PCAP_TIMEOUT to 1, so Kali is able to scan again
I came across this problem by testing the tool on an (currently) up to date Kali. The programm runs fine without any error messages, but still does not find any devices on the network.

After a few hours I eventually found out, that the function `pcap_next_ex` never returned a packet, even if Wireshark (or tcpdump) showed one incoming. Changing the pcap timeout to 1 solved the problem.

I suspect this might be an issue because with timeout 0 the operating system waits until the buffer is filled up. Therefore, on a system with very little network traffic (e.g. a Kali doing only a litte scan) the buffer fills not up until the timeout implemented in the scan tool directly is reached, meaning that the scan tool does not receive any packets in this time, thus finding no hosts at all.

This fix is not very sophisticated and could cause problems with other operating systems. I am not sure why the differentiation between different OSes was made in the first place.

Additionally to Kali I tested this patch with Ubuntu 18.04LTS.