Add MLD(v1) Query tool mldq6
This adds the new "mldq6" tool. It allows sending Query messages for the
Multicast Listener Discovery protocol.

MLD is a mandatory part of the IPv6 standard. So if a network involves
multicast snooping switches/bridges or IPv6 multicast routers malicious
or broken MLD Query messages, especially if an IPv6 implementations is
not strictly validating MLD Query messages, can disrupt the network.

The code was copied and adapted from the rs6 tool, as the MLD Query
protocol is similarly simple.

For now this only allows creating MLDv1 Queries (RFC2710). MLDv2 Queries
(RFC3810) is still ToDo.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>