Improve script6 get-suffix-for-tld
You can now specify a DNS zone, and script6 will only print domainnames
for that zone. e.g.
   script6 get-suffix-for-tld gov.au

will print
   XXXX.gov.au

but not
   XXXX.edu.au

Additionally, you can specify a second (optional) numerical argument,
that requires a specific number of DNS labels below the specified zone, as
in e.g.:
   script6 get-suffix-for-tld gov.au 1

This will print
   XXXX.gov.au

but not
   misc.XXXX.gov.au