Change all size_t nw to int to fix -Wsign-compare
Trying to remove comparisons between signed and unsigned integers,
this is low-hanging fruit.

nw is declared as size_t but it is used to receive the result from
pcap_inject which is defined as int, as the return code in an error
situation is -1.

It did not really affect the call to printf() as it was casted as
long unsigned int. My guess is that the case was just to avoid using
%zu. If so, then the call to printf() must use the new type. Also,
if any future negative value is added to pcap_inject() it will be
properly reflected in printf() and not lost because of being treated
as unsigned.